remove testing
We may consider making a dedicated pytest plugin at some point in the future
